+++ app/web/wayforge-client/src/services/reaction.ts
@@ -17,7 +17,7 @@ import {
 import type { Identified } from "~/packages/anvl/src/id/identified"
 import { now } from "~/packages/anvl/src/id/now"
 import type { Json } from "~/packages/anvl/src/json"
-import type { JsonSchema } from "~/packages/anvl/src/json/json-schema/json-schema"
+import type { JsonSchema } from "~/packages/anvl/src/json-schema/json-schema"
 import { removeFromIndex } from "~/packages/hamr/recoil-tools/recoil-index"
 import type { Transact } from "~/packages/hamr/recoil-tools/recoil-utils"
