+++ packages/anvl/package.json
@@ -1,12 +1,19 @@
 {
   "name": "anvl",
   "private": true,
-  "version": "0.0.0",
+  "version": "0.1.0",
   "description": "core typescript utilities for wayforge",
   "main": "~/lib/index.js",
   "types": "~/lib/index.d.ts",
   "files": [
-    "**/*"
+    "dist",
+    "array",
+    "function",
+    "id",
+    "join",
+    "json",
+    "nullish",
+    "number"
   ],
   "scripts": {
     "dev": "cd example && vite",
@@ -15,17 +22,17 @@
     "test": "vitest",
     "test:once": "vitest run"
   },
-  "devDependencies": {
-    "eslint": "8.34.0",
+  "dependencies": {
     "fp-ts": "2.13.1",
-    "io-ts": "2.2.20",
-    "vite": "4.0.4",
+    "io-ts": "2.2.20"
+  },
+  "devDependencies": {
     "vite-tsconfig-paths": "4.0.5",
     "vitest": "0.28.5"
   },
   "peerDependencies": {
-    "fp-ts": "2.13.1",
-    "io-ts": "2.2.20"
+    "fp-ts": "^2.13.1",
+    "io-ts": "^2.2.20"
   },
   "repository": {
     "type": "git",
