+++ packages/anvl/src/json-schema/integer.test.ts
