+++ packages/anvl/src/json-schema/integer.ts
