+++ packages/anvl/src/json-schema/validate-instance-tests/logical-conditional.test.ts
@@ -1,4 +1,4 @@
-import { select } from "../../../object"
+import { select } from "../../object"
 import type {
   BooleanSchema,
   ConditionalSchema,
