+++ packages/anvl/src/json-schema/validate-instance-tests/logical-exclusive.test.ts
@@ -1,4 +1,4 @@
-import { select } from "../../../object"
+import { select } from "../../object"
 import { Int } from "../integer"
 import type {
   ExclusiveSchema,
