+++ packages/anvl/src/json-schema/validate-instance-tests/logical-intersection.test.ts
@@ -1,6 +1,10 @@
-import { select } from "../../../object"
+import { select } from "../../object"
 import { Int } from "../integer"
-import type { IntegerSchema, NumberSchema, UnionSchema } from "../json-schema"
+import type {
+  IntegerSchema,
+  IntersectionSchema,
+  NumberSchema,
+} from "../json-schema"
 import { validateBy } from "../validate-instance"
 
 const integerFrom2To8: IntegerSchema = {
@@ -12,15 +16,25 @@ const number123: NumberSchema = {
   type: `number`,
   enum: [1, 2, 3],
 }
-const SCHEMA: UnionSchema = {
-  anyOf: [integerFrom2To8, number123],
+const SCHEMA: IntersectionSchema = {
+  allOf: [integerFrom2To8, number123],
 }
 const validate = validateBy(SCHEMA)
 
-describe(`union validation`, () => {
-  it(`validates an instance that validates against one option`, () => {
+describe(`intersection validation`, () => {
+  it(`does not pass an instance that fails against one option`, () => {
     const instance = Int(4)
-    expect(validate(instance)).toStrictEqual({ isValid: true, violations: [] })
+    expect(validate(instance)).toStrictEqual({
+      isValid: false,
+      violations: [
+        {
+          instance,
+          schema: {
+            allOf: [select(`enum`)(number123)],
+          },
+        },
+      ],
+    })
   })
   it(`validates an instance that validates against more than one option`, () => {
     const instance = Int(2)
@@ -34,7 +48,7 @@ describe(`union validation`, () => {
         {
           instance,
           schema: {
-            anyOf: [
+            allOf: [
               select(`maximum`)(integerFrom2To8),
               select(`enum`)(number123),
             ],
