+++ packages/anvl/src/json-schema/validate-instance-tests/logical-negation.test.ts
