+++ packages/anvl/src/json-schema/validate-instance-tests/logical-union.test.ts
@@ -1,10 +1,6 @@
-import { select } from "../../../object"
+import { select } from "../../object"
 import { Int } from "../integer"
-import type {
-  IntegerSchema,
-  IntersectionSchema,
-  NumberSchema,
-} from "../json-schema"
+import type { IntegerSchema, NumberSchema, UnionSchema } from "../json-schema"
 import { validateBy } from "../validate-instance"
 
 const integerFrom2To8: IntegerSchema = {
@@ -16,25 +12,15 @@ const number123: NumberSchema = {
   type: `number`,
   enum: [1, 2, 3],
 }
-const SCHEMA: IntersectionSchema = {
-  allOf: [integerFrom2To8, number123],
+const SCHEMA: UnionSchema = {
+  anyOf: [integerFrom2To8, number123],
 }
 const validate = validateBy(SCHEMA)
 
-describe(`intersection validation`, () => {
-  it(`does not pass an instance that fails against one option`, () => {
+describe(`union validation`, () => {
+  it(`validates an instance that validates against one option`, () => {
     const instance = Int(4)
-    expect(validate(instance)).toStrictEqual({
-      isValid: false,
-      violations: [
-        {
-          instance,
-          schema: {
-            allOf: [select(`enum`)(number123)],
-          },
-        },
-      ],
-    })
+    expect(validate(instance)).toStrictEqual({ isValid: true, violations: [] })
   })
   it(`validates an instance that validates against more than one option`, () => {
     const instance = Int(2)
@@ -48,7 +34,7 @@ describe(`intersection validation`, () => {
         {
           instance,
           schema: {
-            allOf: [
+            anyOf: [
               select(`maximum`)(integerFrom2To8),
               select(`enum`)(number123),
             ],
