+++ packages/anvl/src/number/index.ts
@@ -1,0 +1,2 @@
+export * from "./clamp"
+export * from "./wrap"
