+++ packages/anvl/src/string/index.ts
@@ -1,0 +1,1 @@
+export * from "./split"
