+++ packages/hamr/react-ui/json-editor/developer-interface.tsx
@@ -6,7 +6,7 @@ import Ajv from "ajv"
 import type { SetterOrUpdater } from "recoil"
 
 import type { Json, JsonTypes } from "~/packages/anvl/src/json"
-import type { JsonSchema } from "~/packages/anvl/src/json/json-schema/json-schema"
+import type { JsonSchema } from "~/packages/anvl/src/json-schema/json-schema"
 
 import type { JsxElements } from "."
 import type { JsonEditorComponents } from "./default-components"
