+++ packages/hamr/react-ui/json-editor/json-editor-internal.tsx
@@ -5,8 +5,8 @@ import type { SetterOrUpdater } from "recoil"
 
 import { doNothing } from "~/packages/anvl/src/function"
 import type { Json, JsonTypes } from "~/packages/anvl/src/json"
-import type { JsonSchema } from "~/packages/anvl/src/json/json-schema/json-schema"
 import { refineJsonType } from "~/packages/anvl/src/json/refine"
+import type { JsonSchema } from "~/packages/anvl/src/json-schema/json-schema"
 
 import type { JsxElements } from "."
 import { SubEditors } from "."
